+++ RELEASE.md
@@ -1,0 +1,3 @@
+Release type: patch
+
+This release moves qtgql subcomponents into namespaces.
