+++ scripts/create_new_submodule.py
@@ -7,7 +7,10 @@ QTGQL = ROOT_DIR / "qtgql"
 def main(name: str):
     submod = QTGQL / name
     submod.mkdir(exist_ok=True)
-    (submod / "inc" / "qtgql" / name).mkdir(parents=True)
+    inc_dir = submod / "inc" / "qtgql" / name
+    inc_dir.mkdir(parents=True, exist_ok=True)
+    (inc_dir / f"{name}.hpp").write_text("# pragma once")
 
 
-main("gqlwstransport")
+if __name__ == "__main__":
+    main(input("enter component name"))
